runtimetest: Make TAP output more granular
Granular TAP output makes it easier to see what's being tested (and
what's going wrong).  You could feed this into any TAP harness you
like, but I've chosen prove (it seems popular).  You need a TAP
harness looking for test failures, because runtimetest now returns
zero when it successfully runs the tests (regardless of whether the
tests all pass).

With this commit, runtimetest will exit zero when it can successfully
*run the tests* (regardless of whether all the tests pass).  A TAP
harness like prove should be used to consume the output TAP to decide
if validation succeeded or not.  To avoid confusing the TAP consumer,
all of the non-TAP logging goes to stderr (and not stdout).

Signed-off-by: W. Trevor King <wking@tremily.us>